添加自发光

+++ gpu-version/texture.cuh
@@ -5,7 +5,7 @@
 
 class mytexture {
 public:
-    __device__ virtual color value(double u, double v, const point3 &p) const = 0;
+    __device__ virtual color value(float u, float v, const point3 &p) const = 0;
 };
 
 class solid_color : public mytexture {
@@ -14,9 +14,9 @@ public:
 
     __device__ solid_color(color c) : color_value(c) {}
 
-    __device__ solid_color(double red, double green, double blue) : solid_color(color(red, green, blue)) {}
+    __device__ solid_color(float red, float green, float blue) : solid_color(color(red, green, blue)) {}
 
-    __device__ virtual color value(double u, double b, const vec3 &p) const override {
+    __device__ virtual color value(float u, float b, const vec3 &p) const override {
         return color_value;
     }
 
@@ -34,7 +34,7 @@ public:
     __device__ checker_texture(color c1, color c2) : even(new solid_color(c1)),
                                                      odd(new solid_color(c2)) {}
 
-    __device__ virtual color value(double u, double v, const point3 &p) const override {
+    __device__ virtual color value(float u, float v, const point3 &p) const override {
         auto sines = sin(10 * p.x()) * sin(10 * p.y()) * sin(10 * p.z());
 
         if (sines < 0)
